# Cache commit characters raw json when serializing completion lists.