### Before
|                                              Method |      Mean |     Error |    StdDev |   Op/s |     Gen 0 |     Gen 1 |    Gen 2 | Allocated |
|---------------------------------------------------- |----------:|----------:|----------:|-------:|----------:|----------:|---------:|----------:|
| 'Component Completion List Roundtrip Serialization' | 32.623 ms | 0.6457 ms | 1.1139 ms |  30.65 | 2125.0000 | 1125.0000 | 187.5000 |  13.66 MB |
|           'Component Completion List Serialization' |  8.734 ms | 0.1084 ms | 0.1014 ms | 114.50 |  312.5000 |  234.3750 | 234.3750 |   2.42 MB |

### After
|                                              Method |      Mean |     Error |    StdDev |   Op/s |     Gen 0 |     Gen 1 |    Gen 2 | Allocated |
|---------------------------------------------------- |----------:|----------:|----------:|-------:|----------:|----------:|---------:|----------:|
| 'Component Completion List Roundtrip Serialization' | 28.129 ms | 0.3938 ms | 0.3684 ms |  35.55 | 2156.2500 | 1187.5000 | 218.7500 |   13.6 MB |
|           'Component Completion List Serialization' |  6.081 ms | 0.0450 ms | 0.0399 ms | 164.46 |  312.5000 |  242.1875 | 226.5625 |    2.4 MB |


**Roundtrip op/s improvement:** 1.15x faster
**Serialization op/s improvement:** 1.44x faster


Part of dotnet/aspnetcore#29017